Move components to a directory, create utils/constants
