+++ src/components/HotspotInfo/HotspotInfo.js
