+++ src/components/HotspotInfo/index.js
