+++ src/components/Hotspots/Hotspots.js
