+++ src/components/Legend/index.js
