+++ src/components/RewardLegend/RewardLegend.js
