+++ src/components/RewardLegend/index.js
