+++ src/constants/index.js
@@ -1,0 +1,3 @@
+import { MapConstants } from './MapConstants';
+
+export default MapConstants;
